ITC-15 works suprisingly well, added some forms
