+++ config/sync/core.entity_form_display.node.event.default.yml
@@ -192,7 +192,7 @@ content:
     type: options_buttons
     region: content
   field_event_address:
-    weight: 19
+    weight: 20
     settings: {  }
     third_party_settings: {  }
     type: address_default
@@ -220,7 +220,7 @@ content:
     settings:
       display_label: true
     type: boolean_checkbox
-    weight: 20
+    weight: 21
     third_party_settings: {  }
   field_event_enroll_outside_group:
     weight: 123
@@ -246,7 +246,7 @@ content:
     type: image_widget_crop
     region: content
   field_event_location:
-    weight: 18
+    weight: 19
     settings:
       size: 60
       placeholder: ''
@@ -264,7 +264,7 @@ content:
     third_party_settings: {  }
     region: content
   field_event_with_products:
-    weight: 14
+    weight: 18
     settings:
       display_label: true
     third_party_settings: {  }
@@ -280,7 +280,7 @@ content:
   groups:
     type: social_group_selector_widget
     settings: {  }
-    weight: 16
+    weight: 23
     region: content
     third_party_settings: {  }
   langcode:
