+++ html/modules/custom/itchefer_event_enrollment/itchefer_event_enrollment.routing.yml
@@ -1,7 +1,13 @@
-itchefer_event_enrollment.content:
-  path: '/hello'
+itchefer_event_enrollment.product_modal_form:
+  path: '/node/{node}/product-modal-form'
   defaults:
-    _controller: '\Drupal\itchefer_event_enrollment\Controller\EventEnrollmentController::content'
-    _title: 'Hello World'
+    _controller: '\Drupal\itchefer_event_enrollment\Controller\EventEnrollmentController::productModalDialog'
+    _title_callback: '\Drupal\social_event\Controller\EnrollRequestDialogController::enrollTitle'
+  options:
+    parameters:
+      node:
+        type: entity:node
   requirements:
-    _permission: 'access content'
+    _custom_access: '\Drupal\social_event\Controller\EnrollRequestDialogController::enrollAccess'
+
+
